+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased] - TBD
 
+## [1.5.0] - 2026-02-05
+### Added
+- Event Subscriptions & Webhooks:
+    - Get Callbacks History by ID: webhook callback event history related to subscription ID with advanced filtering and sorting capabilities
+
 ## [1.4.0] - 2026-01-13
 ### Added
 - Event Subscriptions & Webhooks: